fix(layout): 修改layout鉴权逻辑，抽离匹配方法
